refactor: pages to rsc, split client components
