+++ www/lib/client.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { motion } from 'framer-motion';
+
+export const MotionDiv = motion.div;
+export const MotionA = motion.a;
+export const MotionP = motion.p;
