+++ www/lib/components/AnimatedLine.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useRef, useEffect } from 'react';
 
 type AnimatedLineProps = {
